--- core/src/main/java/com/datastrato/gravitino/authorization/AuthorizationConstants.java
@@ -1,10 +1,0 @@
-/*
- * Copyright 2024 Datastrato Pvt Ltd.
- * This software is licensed under the Apache License version 2.
- */
-package com.datastrato.gravitino.authorization;
-
-public class AuthorizationConstants {
-  public static final String SYSTEM_CATALOG_RESERVED_NAME = "system";
-  public static final String USER_SCHEMA_NAME = "user";
-}
