+++ core/src/main/java/com/datastrato/gravitino/meta/CatalogEntity.java
@@ -24,6 +24,8 @@ import lombok.ToString;
 @ToString
 public class CatalogEntity implements Entity, Auditable, HasIdentifier {
 
+  public static final String SYSTEM_CATALOG_RESERVED_NAME = "system";
+
   public static final Field ID =
       Field.required("id", Long.class, "The catalog's unique identifier");
   public static final Field NAME = Field.required("name", String.class, "The catalog's name");
