+++ core/src/main/java/com/datastrato/gravitino/meta/UserEntity.java
@@ -21,6 +21,8 @@ import lombok.ToString;
 @ToString
 public class UserEntity implements User, Entity, Auditable, HasIdentifier {
 
+  public static final String USER_SCHEMA_NAME = "user";
+
   public static final Field ID =
       Field.required("id", Long.class, " The unique id of the user entity.");
